Build basic functionality and start write up of model(#3)
* Add numpy to requirements.

* Write test for compute probability

for trivial fitness function.

Unsure on correct answer to prob.

* Added generate_transition_matrix function

and tests. Updated requirements to include

sympy.

* Added generate_transition_matrix and related

tests. Updated requirements. Ran black on this one

* Update tex/main.tex

Co-authored-by: Vince Knight <vince@vknight.org>

* Updated with feedback from the pull request.

Currently throws up an error when trying to

update the transition matrix with symbolic values

* Updated with the feedback from the last PR. Throws

up an error when trying to overwrite the

transition matrix with symbolic values.

* Updated the main.tex file with mathematical

details and section headers.

* Updated main.tex wording and added Vince's

bugfix for the previous issue. It does work,

and the test_main file fails properly. Still

needs a few docstrings.

* Ran black on previous push

* Updated main.tex with more detail and added

the pdf.

* Updated main.tex and main.pdf with better

wording

* Added Kwargs functionality and tests

* Added docstrings and organised test_main.py

* Ran black on src files

* Updated main.tex and made main.pdf. Also provided

a bugfix in text_main.py

---------

Co-authored-by: Vince Knight <vince@vknight.org>