Add README to src
I believe we agreed on the name Maynard for the
library. For now, I have referred to the code as
Maynard, but in the sections of code I have
written import main rather than import maynard
as this is how it functions at the moment. This
can be changed further on down the line.